fix: [MED-55] При введении количества лекарств с запятой, сохраняется как 0
Добавил расширение, для преобразования текста в число
Изменил значение в базе данных с Float на Double. Это временное решение. До выяснения в каком формате будут хранится данные.

+++ MedicinesV2/Models/CoreData/Medicine+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension Medicine {
         return NSFetchRequest<Medicine>(entityName: "Medicine")
     }
 
-    @NSManaged public var amount: Float
+    @NSManaged public var amount: Double
     @NSManaged public var title: String?
     @NSManaged public var type: String?
     @NSManaged public var expiryDate: Date?
